add functional to posts: add, update, delete

+++ src/main/java/com/projects/notasaint/socialmediaRESTAPI/repositories/PostRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post, Integer> {
-    Optional<Post> findPostById(int id);
+    Optional<Post> findPostById(Long id);
 }
